Fix bug where the parser wouldn't assign the default value of an option
The line which assigns the default was buried inside two loops that iterate
over the arguments and options that matched each specific argument. Default
values would only be assigned if the option was present in the arguments.

Fixed by adding a separate loop that initializes the values with defaults
before parsing begins.